move resolve-input-dir
