+++ src/resolve-input-directory.js
@@ -1,0 +1,18 @@
+const broccoliNodeInfo    = require('broccoli-node-info');
+
+module.exports = function resolveInputDirectory(inputNodes) {
+  if (typeof inputNodes === 'string') {
+    return inputNodes;
+  }
+
+  const nodeInfo = broccoliNodeInfo.getNodeInfo(inputNodes);
+  if (nodeInfo.nodeType === 'source') {
+    return nodeInfo.sourceDirectory;
+  }
+
+  if (nodeInfo.inputNodes.length > 1) {
+    throw new Error('broccoli-stylelint can only handle one:* broccoli nodes, but part of the given input pipeline is a many:* node. (broccoli-merge-trees is an example of a many:* node) Please perform many:* operations after linting.');
+  }
+
+  return resolveInputDirectory(nodeInfo.inputNodes[0]);
+}
